refactor: move type home.ts under screen folder
